gitserver: Move VCSSyncer factory to vcssyncer package
Just moving some code around to colocate things of similar concerns more closely.
