+++ cmd/gitserver/internal/vcssyncer/BUILD.bazel
@@ -26,24 +26,30 @@ go_library(
         "//cmd/gitserver/internal/gitserverfs",
         "//cmd/gitserver/internal/perforce",
         "//cmd/gitserver/internal/urlredactor",
+        "//internal/actor",
         "//internal/api",
         "//internal/codeintel/dependencies",
         "//internal/conf",
         "//internal/conf/reposource",
+        "//internal/database",
         "//internal/env",
         "//internal/errcode",
+        "//internal/extsvc",
         "//internal/extsvc/crates",
         "//internal/extsvc/gomodproxy",
         "//internal/extsvc/jvmpackages/coursier",
         "//internal/extsvc/npm",
         "//internal/extsvc/pypi",
         "//internal/extsvc/rubygems",
+        "//internal/httpcli",
+        "//internal/jsonc",
         "//internal/observation",
         "//internal/unpack",
         "//internal/vcs",
         "//internal/wrexec",
         "//lib/errors",
         "//schema",
+        "@com_github_json_iterator_go//:go",
         "@com_github_sourcegraph_log//:log",
         "@org_golang_x_mod//module",
         "@org_golang_x_mod//zip",
@@ -60,6 +66,7 @@ go_test(
         "vcs_syncer_npm_packages_test.go",
         "vcs_syncer_perforce_test.go",
         "vcs_syncer_python_packages_test.go",
+        "vcs_syncer_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":vcssyncer"],
@@ -76,7 +83,9 @@ go_test(
         "//internal/codeintel/dependencies",
         "//internal/conf/reposource",
         "//internal/database",
+        "//internal/database/dbmocks",
         "//internal/database/dbtest",
+        "//internal/extsvc",
         "//internal/extsvc/jvmpackages/coursier",
         "//internal/extsvc/npm",
         "//internal/extsvc/npm/npmtest",
@@ -86,6 +95,7 @@ go_test(
         "//internal/observation",
         "//internal/ratelimit",
         "//internal/testutil",
+        "//internal/types",
         "//internal/vcs",
         "//lib/errors",
         "//schema",
