+++ cmd/gitserver/internal/vcssyncer/vcs_syncer.go
@@ -4,9 +4,26 @@ import (
 	"context"
 	"os/exec"
 
+	jsoniter "github.com/json-iterator/go"
+
 	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/common"
+	"github.com/sourcegraph/sourcegraph/cmd/gitserver/internal/gitserverfs"
+	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies"
+	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/crates"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/gomodproxy"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/npm"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/pypi"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/rubygems"
+	"github.com/sourcegraph/sourcegraph/internal/httpcli"
+	"github.com/sourcegraph/sourcegraph/internal/jsonc"
 	"github.com/sourcegraph/sourcegraph/internal/vcs"
+	"github.com/sourcegraph/sourcegraph/internal/wrexec"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
+	"github.com/sourcegraph/sourcegraph/schema"
 )
 
 // VCSSyncer describes whether and how to sync content from a VCS remote to
@@ -29,6 +46,122 @@ type VCSSyncer interface {
 	Fetch(ctx context.Context, remoteURL *vcs.URL, repoName api.RepoName, dir common.GitDir, revspec string) ([]byte, error)
 	// RemoteShowCommand returns the command to be executed for showing remote.
 	RemoteShowCommand(ctx context.Context, remoteURL *vcs.URL) (cmd *exec.Cmd, err error)
+}
+
+type NewVCSSyncerOpts struct {
+	ExternalServiceStore    database.ExternalServiceStore
+	RepoStore               database.RepoStore
+	DepsSvc                 *dependencies.Service
+	Repo                    api.RepoName
+	ReposDir                string
+	CoursierCacheDir        string
+	RecordingCommandFactory *wrexec.RecordingCommandFactory
+}
+
+func NewVCSSyncer(ctx context.Context, opts *NewVCSSyncerOpts) (VCSSyncer, error) {
+	// We need an internal actor in case we are trying to access a private repo. We
+	// only need access in order to find out the type of code host we're using, so
+	// it's safe.
+	r, err := opts.RepoStore.GetByName(actor.WithInternalActor(ctx), opts.Repo)
+	if err != nil {
+		return nil, errors.Wrap(err, "get repository")
+	}
+
+	extractOptions := func(connection any) (string, error) {
+		for _, info := range r.Sources {
+			extSvc, err := opts.ExternalServiceStore.GetByID(ctx, info.ExternalServiceID())
+			if err != nil {
+				return "", errors.Wrap(err, "get external service")
+			}
+			rawConfig, err := extSvc.Config.Decrypt(ctx)
+			if err != nil {
+				return "", err
+			}
+			normalized, err := jsonc.Parse(rawConfig)
+			if err != nil {
+				return "", errors.Wrap(err, "normalize JSON")
+			}
+			if err = jsoniter.Unmarshal(normalized, connection); err != nil {
+				return "", errors.Wrap(err, "unmarshal JSON")
+			}
+			return extSvc.URN(), nil
+		}
+		return "", errors.Errorf("unexpected empty Sources map in %v", r)
+	}
+
+	switch r.ExternalRepo.ServiceType {
+	case extsvc.TypePerforce:
+		var c schema.PerforceConnection
+		if _, err := extractOptions(&c); err != nil {
+			return nil, err
+		}
+
+		p4Home, err := gitserverfs.MakeP4HomeDir(opts.ReposDir)
+		if err != nil {
+			return nil, err
+		}
+
+		return NewPerforceDepotSyncer(&c, p4Home), nil
+	case extsvc.TypeJVMPackages:
+		var c schema.JVMPackagesConnection
+		if _, err := extractOptions(&c); err != nil {
+			return nil, err
+		}
+		return NewJVMPackagesSyncer(&c, opts.DepsSvc, opts.CoursierCacheDir), nil
+	case extsvc.TypeNpmPackages:
+		var c schema.NpmPackagesConnection
+		urn, err := extractOptions(&c)
+		if err != nil {
+			return nil, err
+		}
+		cli, err := npm.NewHTTPClient(urn, c.Registry, c.Credentials, httpcli.ExternalClientFactory)
+		if err != nil {
+			return nil, err
+		}
+		return NewNpmPackagesSyncer(c, opts.DepsSvc, cli), nil
+	case extsvc.TypeGoModules:
+		var c schema.GoModulesConnection
+		urn, err := extractOptions(&c)
+		if err != nil {
+			return nil, err
+		}
+		cli := gomodproxy.NewClient(urn, c.Urls, httpcli.ExternalClientFactory)
+		return NewGoModulesSyncer(&c, opts.DepsSvc, cli), nil
+	case extsvc.TypePythonPackages:
+		var c schema.PythonPackagesConnection
+		urn, err := extractOptions(&c)
+		if err != nil {
+			return nil, err
+		}
+		cli, err := pypi.NewClient(urn, c.Urls, httpcli.ExternalClientFactory)
+		if err != nil {
+			return nil, err
+		}
+		return NewPythonPackagesSyncer(&c, opts.DepsSvc, cli, opts.ReposDir), nil
+	case extsvc.TypeRustPackages:
+		var c schema.RustPackagesConnection
+		urn, err := extractOptions(&c)
+		if err != nil {
+			return nil, err
+		}
+		cli, err := crates.NewClient(urn, httpcli.ExternalClientFactory)
+		if err != nil {
+			return nil, err
+		}
+		return NewRustPackagesSyncer(&c, opts.DepsSvc, cli), nil
+	case extsvc.TypeRubyPackages:
+		var c schema.RubyPackagesConnection
+		urn, err := extractOptions(&c)
+		if err != nil {
+			return nil, err
+		}
+		cli, err := rubygems.NewClient(urn, c.Repository, httpcli.ExternalClientFactory)
+		if err != nil {
+			return nil, err
+		}
+		return NewRubyPackagesSyncer(&c, opts.DepsSvc, cli), nil
+	}
+	return NewGitRepoSyncer(opts.RecordingCommandFactory), nil
 }
 
 type notFoundError struct{ error }
