+++ cmd/gitserver/shared/BUILD.bazel
@@ -31,22 +31,14 @@ go_library(
         "//internal/debugserver",
         "//internal/encryption/keyring",
         "//internal/env",
-        "//internal/extsvc",
-        "//internal/extsvc/crates",
-        "//internal/extsvc/gomodproxy",
-        "//internal/extsvc/npm",
-        "//internal/extsvc/pypi",
-        "//internal/extsvc/rubygems",
         "//internal/gitserver/v1:gitserver",
         "//internal/goroutine",
         "//internal/goroutine/recorder",
         "//internal/grpc",
         "//internal/grpc/defaults",
         "//internal/hostname",
-        "//internal/httpcli",
         "//internal/httpserver",
         "//internal/instrumentation",
-        "//internal/jsonc",
         "//internal/observation",
         "//internal/ratelimit",
         "//internal/requestclient",
@@ -54,8 +46,6 @@ go_library(
         "//internal/trace",
         "//internal/wrexec",
         "//lib/errors",
-        "//schema",
-        "@com_github_json_iterator_go//:go",
         "@com_github_sourcegraph_log//:log",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_x_sync//semaphore",
@@ -74,15 +64,9 @@ go_test(
     # path is sandboxed properly.
     env = {"COURSIER_CACHE_DIR": "/tmp"},
     deps = [
-        "//internal/api",
-        "//internal/codeintel/dependencies",
-        "//internal/database/dbmocks",
-        "//internal/extsvc",
-        "//internal/types",
         "@com_github_google_go_cmp//cmp",
         "@com_github_sourcegraph_log//:log",
         "@com_github_sourcegraph_log//logtest",
-        "@com_github_stretchr_testify//require",
         "@org_golang_google_grpc//:go_default_library",
     ],
 )
